+++ lib/src/dto/remote_configuration_assignment_type.dart
@@ -1,0 +1,15 @@
+import 'package:json_annotation/json_annotation.dart';
+
+enum QRemoteConfigurationAssignmentType {
+  /// Unknown assignment type
+  @JsonValue('unknown')
+  unknown,
+
+  /// Auto
+  @JsonValue('auto')
+  auto,
+
+  /// Manual
+  @JsonValue('manual')
+  manual,
+}